+++ src/components/common/input/Input.tsx
@@ -1,8 +1,26 @@
 'use client';
 
+export type InputType = 'box' | 'textarea' | 'line';
+
+interface BaseProps {
+  label?: string;
+  placeholder?: string;
+  box?: InputType;
+  value?: string;
+  onChange?: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
+  ) => void;
+}
+
+export type InputProps = BaseProps &
+  (
+    | (React.InputHTMLAttributes<HTMLInputElement> & { box?: 'box' | 'line' })
+    | (React.TextareaHTMLAttributes<HTMLTextAreaElement> & { box: 'textarea' })
+  );
+
 // components/CommonInput.tsx
 import React from 'react';
-import { InputProps } from '@/components/common/input/types';
+
 // import classNames from 'classnames';
 
 const CommonInput: React.FC<InputProps> = ({
